first version2.0

+++ internal/models/person.go
@@ -1,9 +1,9 @@
 package models
 
 type Person struct {
-	ID      int
+	ID      int64
 	Name    string
-	Age     int
+	Age     int64
 	Address string
 	Work    string
 }
